feat(website): move creator details into a "More details" button
The creator details modal was opened via a small info icon next to the
"Details" section header. Move it to be a "More details" button placed
next to the "Export / Cite" button in the SeqSet actions toolbar, where
it is more discoverable and consistent with the other actions.

The modal content (the clarifying note plus the "Created by" and
"Created date" fields) now lives in SeqSetItemActions, which receives
the seqSetAuthor prop. SeqSetItem no longer needs the author prop or the
modal state.

The seqset integration test page object is updated to click the new
"More details" button.
