+++ integration-tests/tests/pages/seqset.page.ts
@@ -83,7 +83,7 @@ export class SeqSetPage {
 
     async expectCreatorDetailsInModal() {
         await expect(this.page.getByText('Created by', { exact: true })).toBeHidden();
-        await this.page.getByRole('button', { name: 'Show creator details' }).click();
+        await this.page.getByRole('button', { name: 'More details' }).click();
         await expect(this.page.getByRole('heading', { name: 'Creator details' })).toBeVisible();
         await expect(this.page.getByText('Created by', { exact: true })).toBeVisible();
         await expect(this.page.getByText('Created date', { exact: true })).toBeVisible();
