+++ website/src/pages/seqsets/[seqSetId].[version].astro
@@ -110,6 +110,7 @@ const secondaryErrors = (
                     clientConfig={clientConfig}
                     accessToken={accessToken}
                     seqSet={seqSetResponse.value}
+                    seqSetAuthor={seqSetAuthor}
                     seqSetRecords={seqSetRecordsResponse.value}
                     isAdminView={seqSetResponse.value.createdBy === username}
                     databaseName={websiteConfig.name}
@@ -120,7 +121,6 @@ const secondaryErrors = (
                     accessToken={accessToken}
                     seqSetAccessionVersion={seqSetAccessionVersion}
                     seqSet={seqSetResponse.value}
-                    seqSetAuthor={seqSetAuthor}
                     seqSetRecords={seqSetRecordsResponse.value}
                     citedByData={citedByData}
                     seqSetGraphs={seqSetGraphs}
